# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Librarr API",
    "description": "Self-hosted book, audiobook, and manga search and download manager — the *arr for books. The HTTP/JSON API powers both the SPA and any external tooling (CLI scripts, AI agents, *arr-family integrations).",
    "version": "1.x",
    "license": {"name": "MIT", "identifier": "MIT"}
  },
  "servers": [
    {"url": "http://localhost:5050", "description": "Default local install"}
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key for programmatic access. Also accepted via ?apikey= query parameter."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "librarr_session",
        "description": "Set by POST /api/login. Used by the SPA."
      }
    },
    "schemas": {
...
      "PaginatedList": {
        "type": "object",
        "description": "Standard list response shape used across /api/library, /api/activity, etc.",
        "properties": {
          "items":  {"type": "array", "items": {"type": "object"}},
          "total":  {"type": "integer"},
          "limit":  {"type": "integer"},
          "offset": {"type": "integer"}
        }
      }
    }
  },
  "security": [{"apiKey": []}, {"session": []}],
  "paths": {
...
    "/api/library": {
      "get": {
        "summary": "List ebooks in the library",
        "parameters": [
          {"name": "limit",  "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 500, "default": 50}},
          {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}},
          {"name": "type",   "in": "query", "schema": {"type": "string", "enum": ["ebook", "audiobook", "manga"]}},
          {"name": "tag",    "in": "query", "schema": {"type": "string"}, "description": "Filter by tag name."}
        ],
        "responses": {
          "200": {
            "description": "Paginated library items",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedList"}}}
          }
        }
      }
    },
...
    "/torznab/api": {
      "get": {
        "summary": "Torznab indexer API",
        "description": "Lets Prowlarr / Readarr / other *arr apps use Librarr as a Torznab indexer.",
        "parameters": [
          {"name": "t",      "in": "query", "schema": {"type": "string", "enum": ["caps", "search", "book", "audio"]}},
          {"name": "q",      "in": "query", "schema": {"type": "string"}},
          {"name": "title",  "in": "query", "schema": {"type": "string"}},
          {"name": "author", "in": "query", "schema": {"type": "string"}},
          {"name": "apikey", "in": "query", "schema": {"type": "string"}, "description": "Required when TORZNAB_API_KEY is set."}
        ],
        "responses": {
          "200": {"description": "Torznab XML (RSS feed or caps document)", "content": {"application/xml": {}}}
        }
      }
    },
...
  }
}